refactor elevator into folder
