+++ src/main/java/competition/simulation/elevator/ElevatorSimConstants.java
@@ -1,4 +1,4 @@
-package competition.simulation;
+package competition.simulation.elevator;
 
 public class ElevatorSimConstants {
     // These are are all placeholders
@@ -8,5 +8,8 @@ public class ElevatorSimConstants {
     public static final double minElevatorHeightMeters = 0;
     public static final double maxElevatorHeightMeters = 1.25;
     public static final double elevatorBottomSensorTriggerHeight = 0.01;
-    public static final double rotationsPerMeterHeight = 1;
+    public static final double rotationsPerMeterHeight = 1923; // arbitrary big number
+    // this is random to mimic the motor not being zeroed when the robot turns on so the
+    // code will have to handle the calibration thereof
+    public static final double rotationsAtZero = Math.random() * 1000;
 }
